[libc] Added static casts to fix implicit conversion warnings in 32-bit systems
This patch fixes:
* randomness.h and getauxval.cpp were passing ssize_t as size_t
* kernel_statx.h was assigning an uint64_t to uintptr_t
* fopencookie.cpp was trying to create a FileIOResult using ssize_t but
  the constructor expected a size_t
* thread.h was trying to call free_stack (which takes a size_t) with an
  unsigned long long. free_stack does the calculations using uintptr_t,
  so I changed the passing values to size_t

+++ libc/src/__support/threads/thread.h
@@ -97,11 +97,11 @@ struct alignas(STACK_ALIGNMENT) ThreadAttributes {
   //          exits. It will clean up the thread resources once the thread
   //          exits.
   cpp::Atomic<uint32_t> detach_state;
-  void *stack;                  // Pointer to the thread stack
-  unsigned long long stacksize; // Size of the stack
-  unsigned long long guardsize; // Guard size on stack
-  uintptr_t tls;                // Address to the thread TLS memory
-  uintptr_t tls_size;           // The size of area pointed to by |tls|.
+  void *stack;               // Pointer to the thread stack
+  size_t stacksize;          // Size of the stack
+  size_t guardsize;          // Guard size on stack
+  uintptr_t tls;             // Address to the thread TLS memory
+  uintptr_t tls_size;        // The size of area pointed to by |tls|.
   unsigned char owned_stack; // Indicates if the thread owns this stack memory
   int tid;
   ThreadStyle style;
